add overrides to sensor and finish sketching basic infrastructure
