+++ src/reconstruction/CMakeLists.txt
@@ -7,7 +7,6 @@ target_sources(
           ${CMAKE_CURRENT_SOURCE_DIR}/projection_interpolators.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/projective_integrator.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/semantic_integrator.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/sensor_overrides.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/tsdf_interpolators.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/volumetric_map.cpp
 )
